Change interface definition
I redesigned the interface definition of UrlHashRouter.

A caller had to used static member functions to activate a route. This
caused a problem, though. It did not allow me to "detect" if the caller
did not define a route for a particular location.

In order to to "detect" a missing route, I have to cache all the routes.
Using a for loop I can enumerate all potential routes, and using the
Elephant in Cairo method determine if a missing route exists.